+++ phpBB2/language/lang_english/lang_main.php
@@ -1499,6 +1499,7 @@ $lang['Moved'] = 'Moved';
 //signature editor
 $lang['sig_description'] = "Edit Signature (<b>Preview included</b>)";
 $lang['sig_edit'] = "Edit Signature";
+$lang['Signature_new_tab'] = 'Opens in a new tab. Unsaved profile changes remain in this tab.';
 $lang['sig_current'] = "Current Signature";
 $lang['sig_none'] = "No Signature available";
 $lang['sig_save'] = "Save";